separate Folding logic into separate project
also a 'shim' for the new interface that probably needs to be in Uno.UI
