+++ WindowManager/Uno.UI.Foldable-orig0test/Uno.UI.Foldable.csproj
@@ -1,0 +1,66 @@
+<Project Sdk="MSBuild.Sdk.Extras" ToolsVersion="15.0">
+  <PropertyGroup>
+    <TargetFrameworks>MonoAndroid11.0</TargetFrameworks>
+    <TargetFrameworksCI>MonoAndroid11.0;MonoAndroid10.0</TargetFrameworksCI>
+    <!--<GeneratePackageOnBuild Condition="'$(Configuration)'=='Release'">true</GeneratePackageOnBuild>-->
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+
+    <PackageId Condition="'$(UNO_UWP_BUILD)'!='true'">Uno.WinUI.Foldable</PackageId>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(UnoTargetFrameworkOverride)'!=''">
+    <TargetFrameworks>$(UnoTargetFrameworkOverride)</TargetFrameworks>
+  </PropertyGroup>
+
+  <PropertyGroup>
+    <Authors>nventive</Authors>
+    <PackageProjectUrl>https://github.com/unoplatform/uno</PackageProjectUrl>
+    <PackageIconUrl>https://nv-assets.azurewebsites.net/logos/uno.png</PackageIconUrl>
+    <RepositoryUrl>https://github.com/unoplatform/uno</RepositoryUrl>
+    <Description>This package provides the ability for Uno Platform based apps to use dual-screen and foldable devices such as the Surface Duo or others that support Android Jetpack Window Manager</Description>
+    <Copyright>Copyright (C) 2015-2020 nventive inc. - all rights reserved</Copyright>
+  </PropertyGroup>
+
+  <PropertyGroup>
+    <NoWarn>$(NoWarn);NU1701</NoWarn>
+
+    <ProduceReferenceAssembly>false</ProduceReferenceAssembly>
+    <Deterministic>true</Deterministic>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Uno.SourceGenerationTasks" />
+    <PackageReference Include="Uno.Core" />
+    <PackageReference Include="Uno.Core.Build" />
+    <PackageReference Include="Uno.UI" />
+  </ItemGroup>
+
+  <ItemGroup Condition=" '$(TargetFramework)' == 'MonoAndroid11.0' or '$(TargetFramework)' == 'MonoAndroid10.0'">
+    <PackageReference Include="Xamarin.AndroidX.Window.WindowJava" />
+  </ItemGroup>
+
+  <ItemGroup Condition="$(IsMonoAndroid) or $(IsXamarinIOS) or $(IsXamarinMac)">
+    <Reference Include="System.Numerics" />
+    <Reference Include="System.Numerics.Vectors" />
+    <Reference Include="System.json" />
+    <Reference Include="System.Memory" />
+  </ItemGroup>
+
+  <!--<Target Name="_UnoDualScreenOverrideNuget" AfterTargets="AfterBuild" DependsOnTargets="BuiltProjectOutputGroup" Condition="'$(UnoNugetOverrideVersion)'!=''">
+
+    <PropertyGroup>
+      <_OverrideTargetFramework>$(TargetFramework)</_OverrideTargetFramework>
+      <_TargetNugetFolder>$(USERPROFILE)\.nuget\packages\Uno.UI.DualScreen\$(UnoNugetOverrideVersion)\lib\$(_OverrideTargetFramework)</_TargetNugetFolder>
+    </PropertyGroup>
+    <ItemGroup>
+      <_OutputFiles Include="$(TargetDir)**" />
+    </ItemGroup>
+    <MakeDir Directories="$(_TargetNugetFolder)" />
+
+    <Message Importance="high" Text="OVERRIDING NUGET PACKAGE CACHE: $(_TargetNugetFolder)" />
+
+    <Copy SourceFiles="@(_OutputFiles)" DestinationFiles="@(_OutputFiles->'$(_TargetNugetFolder)\%(RecursiveDir)%(Filename)%(Extension)')" />
+    <Copy SourceFiles="@(_OutputFilesPDB)" DestinationFiles="@(_OutputFilesPDB->'$(_TargetNugetFolder)\%(RecursiveDir)%(Filename).pdb')" />
+  </Target>-->
+</Project>
+
