+++ WindowManager/WindowManager/WindowManager.Droid/Properties/AndroidManifest.xml
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="WindowManager" android:versionCode="1" android:versionName="1.0">
-  <uses-sdk android:minSdkVersion="16" android:targetSdkVersion="29" />
-  <application android:label="Uno Window Manager"></application>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="WindowManager.WindowManager" android:versionCode="1" android:versionName="1.0" android:installLocation="auto">
+	<uses-sdk android:minSdkVersion="25" android:targetSdkVersion="30" />
+	<application android:label="Uno Window Manager"></application>
 </manifest>
