+++ WindowManager/WindowManager/WindowManager.Droid/Properties/AssemblyInfo.cs
@@ -42,6 +42,6 @@ using WindowManager;
 // https://platform.uno/docs/articles/uno-development/api-extensions.html
 // which suggests these are only generated in App.InitializeComponent() when found in a NuGet
 // (and not in the current application, as in this case)...
-[assembly: ApiExtension(typeof(IApplicationViewSpanningRects), typeof(FoldableApplicationViewSpanningRects))]
-[assembly: ApiExtension(typeof(INativeFoldableProvider), typeof(FoldableApplicationViewSpanningRects))]
-[assembly: ApiExtension(typeof(INativeHingeAngleSensor), typeof(FoldableHingeAngleSensor))]
+//[assembly: ApiExtension(typeof(IApplicationViewSpanningRects), typeof(FoldableApplicationViewSpanningRects))]
+//[assembly: ApiExtension(typeof(INativeFoldableProvider), typeof(FoldableApplicationViewSpanningRects))]
+//[assembly: ApiExtension(typeof(INativeHingeAngleSensor), typeof(FoldableHingeAngleSensor))]
