+++ WindowManager/WindowManager/WindowManager.Droid/WindowManager.Droid.csproj
@@ -17,7 +17,7 @@
     <AndroidUseAapt2>true</AndroidUseAapt2>
     <GenerateSerializationAssemblies>Off</GenerateSerializationAssemblies>
     <AndroidUseLatestPlatformSdk>False</AndroidUseLatestPlatformSdk>
-    <TargetFrameworkVersion>v10.0</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v11.0</TargetFrameworkVersion>
     <AndroidManifest>Properties\AndroidManifest.xml</AndroidManifest>
     <AndroidUseIntermediateDesignerFile>True</AndroidUseIntermediateDesignerFile>
     <ResourcesDirectory>..\WindowManager.Shared\Strings</ResourcesDirectory>
@@ -92,6 +92,12 @@
   </ItemGroup>
   <ItemGroup>
     <AndroidResource Include="Resources\values\Styles.xml" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Uno.UI.Foldable\Uno.UI.Foldable.csproj">
+      <Project>{36161BAF-7A30-4E92-9924-7A5F176DAE6D}</Project>
+      <Name>Uno.UI.Foldable</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="..\WindowManager.Shared\WindowManager.Shared.projitems" Label="Shared" />
   <Import Project="$(MSBuildExtensionsPath)\Xamarin\Android\Xamarin.Android.CSharp.targets" />
