+++ WindowManager/WindowManager/WindowManager.Shared/MainPage.xaml
@@ -16,7 +16,7 @@
             <RowDefinition Height="7*" />
         </Grid.RowDefinitions>
 		<TextBlock Grid.Row="0"
-                   Text="Window Manager for Uno" 
+                   Text="Window Manager for Uno b2" 
                    FontSize="30" />
         <Button Grid.Row="1"
                 Click="TwoPaneView_Click">TwoPaneView</Button>
